Rename Rust module to `declarative_asn1`
